refactor(readr): add type/component
