+++ packages/readr/types/component.ts
@@ -1,0 +1,13 @@
+export type ArticleCard = {
+  id?: string
+  href?: string
+  title: string
+  image?: string
+  date?: string
+  readTimeText?: string
+  isReport?: boolean
+}
+
+export type ArticleCardWithIsFeatured = ArticleCard & {
+  isFeatured?: boolean
+}
